+++ src/components/Survey/SurveyStep1.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
-import SelectButton from "./SelectButton";
+import SelectButton from "./../SelectButton";
 import { Link } from "react-router-dom";
-import colors from "../styles/colors";
+import colors from "../../styles/colors";
 import { useState } from "react";
 
 const Wrapper = styled.div`
