+++ src/pages/Category/CategoryList.tsx
@@ -1,12 +1,12 @@
 import styled from "styled-components";
-import Header from "../components/Header";
-import TextIconHeader from "../components/TextIconHeader ";
+import Header from "../../components/Header";
+import TextIconHeader from "../../components/TextIconHeader ";
 import { useNavigate, useParams } from "react-router-dom";
-import categories from "../data/categories.json";
-import colors from "../styles/colors";
+import categories from "../../data/categories.json";
+import colors from "../../styles/colors";
 import { useState } from "react";
 import { VscChevronDown } from "react-icons/vsc";
-import SortedModal from "../components/SortedModal";
+import SortedModal from "../../components/SortedModal";
 
 const Wrap = styled.div`
   display: flex;
